pyproject.toml
[project]
name = "adk-bwai"
version = "0.1.0"
description = "Simple project for showing ADK basics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.5.0",
    "pydantic>=2.11.4",
    "yfinance>=0.2.61",
]
